+++ site/content/release-notes.md
@@ -9,6 +9,7 @@ weight = 155
 ### New Features
 
 * We have added a protocol adapter for allowing [Eclipse Kura](https://www.eclipse.com/kura) gateways to publish *control* and *data* messages to Hono's Telemetry and Event API. See [Kura Adapter]({{< relref "user-guide/kura-adapter.md" >}}) for details.
+* `RegistrationClientImpl` now supports caching of registration assertions received from a Device Registration service. The protocol adapters already make use of this feature  so that they do not need to do a remote service invocation unless a cached assertion has expired. The protocol adapters support two new configuration variables to set the minimum and maximum cache size.
 
 ### Fixes & Enhancements